// frontend/src/types/clips.ts
interface HasId {
  _id: string;
}

export interface MetaMeta {
  created?: string;
  createdBy?: string;
  updated?: string;
  updatedBy?: string;
}
// ...
export interface Annotation extends HasId, MetaMeta {
  type: 'Annotation';
  clip: string;
  element?: string;
  relation: Relation;
  target: string;
  quotes: string;
  date?: string;
  role?: string;
  instrument?: string;
  metaDate?: string;
  timecodeStart?: number;
  timecodeEnd?: number;
  confidence?: 'certain' | 'unsure' | 'maybe' | ''; // tbd
  // review Franziska uncertainty stuff before making decision here
  attribution?: string;
  // attribution?: {
  //   zoteroId: string;
  //   locator?: string;
  //   attributionComment?: string;
  // };
  comment?: string; // we want comment threads, right?
  fromPlatformMetadata?: boolean;
  interpretative?: boolean;
  constitutedBy?: string[]; // annotation ids... autocomplete w relation and target/data
  refersTo?: string; // annotation id ("Der Handshake ist ein symbolischer Akt")
  // flags: Flag[];
  notablyAbsent?: boolean;
}